feat(utilities): support collapsible fields
