+++ src/utilities/index.ts
@@ -1,4 +1,4 @@
-import { Block, CollectionConfig, Field, GlobalConfig } from 'payload/types'
+import { Block, CollapsibleField, CollectionConfig, Field, GlobalConfig } from 'payload/types'
 import deepEqual from 'deep-equal'
 import { FieldWithName } from '../types'
 import { slateToHtml, payloadSlateToDomConfig } from 'slate-serializers'
@@ -24,7 +24,8 @@ export const getLocalizedFields = ({
 }: {
   fields: Field[],
   type?: 'json' | 'html',
-}): any[] => ( fields
+}): any[] => ([
+  ...fields
   // localized or group fields only.
   .filter(field => isLocalizedField(field) || containsNestedFields(field))
   // further filter on CrowdIn field type
@@ -34,8 +35,7 @@ export const getLocalizedFields = ({
     }
     return type ? fieldCrowdinFileType(field as FieldWithName) === type : true
   })
-  
-  // recursion for group field
+  // recursion for group, array and blocks field
   .map(field => {
     if (field.type === 'group' || field.type === 'array') {
       return {
@@ -50,7 +50,7 @@ export const getLocalizedFields = ({
       return {
         ...field,
         blocks: 
-          field.blocks.map(block => {
+          field.blocks.map((block: Block) => {
             return {
               fields: getLocalizedFields({
                 fields: block.fields,
@@ -62,6 +62,24 @@ export const getLocalizedFields = ({
     }
     return field
   })
+  .filter(field => field.type !== 'collapsible'),
+  // recursion for collapsible field - flatten results into the returned array
+  ...getCollapsibleLocalizedFields({ fields, type })
+])
+
+export const getCollapsibleLocalizedFields = ({
+  fields,
+  type,
+}: {
+  fields: Field[],
+  type?: 'json' | 'html',
+}): any[] => (
+  fields
+    .filter(field => field.type === 'collapsible')
+    .flatMap(field => getLocalizedFields({
+      fields: (field as CollapsibleField).fields,
+      type
+    }))
 )
 
 export const getLocalizedRequiredFields = (collection: CollectionConfig, type?: 'json' | 'html'): any[] => {
